+++ crates/nostr/src/event/builder.rs
@@ -1345,13 +1345,14 @@ impl EventBuilder {
     /// <https://github.com/nostr-protocol/nips/blob/master/59.md>
     #[inline]
     #[cfg(all(feature = "std", feature = "nip59"))]
-    pub async fn seal<T>(
+    pub async fn seal<T, R>(
         signer: &T,
         receiver_pubkey: &PublicKey,
-        rumor: EventBuilder,
+        rumor: R,
     ) -> Result<Self, Error>
     where
         T: NostrSigner,
+        R: Into<EventBuilderOrUnsignedEvent>,
     {
         Ok(nip59::make_seal(signer, receiver_pubkey, rumor).await?)
     }
@@ -1400,14 +1401,15 @@ impl EventBuilder {
     /// <https://github.com/nostr-protocol/nips/blob/master/59.md>
     #[inline]
     #[cfg(all(feature = "std", feature = "nip59"))]
-    pub async fn gift_wrap<T, I>(
+    pub async fn gift_wrap<T, R, I>(
         signer: &T,
         receiver: &PublicKey,
-        rumor: EventBuilder,
+        rumor: R,
         extra_tags: I,
     ) -> Result<Event, Error>
     where
         T: NostrSigner,
+        R: Into<EventBuilderOrUnsignedEvent>,
         I: IntoIterator<Item = Tag>,
     {
         let seal: Event = Self::seal(signer, receiver, rumor)