+++ CHANGELOG.md
@@ -74,6 +74,7 @@
 * pool: add `ReceiverStream` ([Yuki Kishimoto])
 * sdk: automatically resend event after NIP-42 authentication ([Yuki Kishimoto])
 * sdk: add `Connection::embedded_tor_with_path` ([Yuki Kishimoto])
+* connect: add `NostrConnect::status` ([Yuki Kishimoto])
 * relay-builder: add NIP42 support ([Yuki Kishimoto])
 * relay-builder: add negentropy support ([Yuki Kishimoto])
 * relay-builder: add read/write policy plugins ([v0l])